+++ tools/public_api_guard/core/core.d.ts
@@ -1,3 +1,7 @@
+export interface AbstractType<T> extends Function {
+    prototype: T;
+}
+
 export interface AfterContentChecked {
     ngAfterContentChecked(): void;
 }